Extract issue grant functions from TokenHandler to TokenService for reuse
